8ms,
BVH nodes are now a struct containing index of left and right nodes, saving an LDG instruction.

Update data types and CUDA options for improved performance

- Modified `metalDir` calculation in `Material.h` to ensure normalization.

- Changed node index types to uint32_t.

+++ src/CUDARender.cu
@@ -58,7 +58,7 @@ __global__ void RenderKernel()
 
 	Vec3 pixelColor(0.0f);
 
-	for (uint16_t s = 0; s < params->m_SamplesPerPixel; s++)
+	for (uint32_t s = 0; s < params->m_SamplesPerPixel; s++)
 	{
 		const float2 uv	 = float2 { (x + RandomFloat(seed)) * params->ResolutionInfo.x, 1.0f - (y + RandomFloat(seed)) * params->ResolutionInfo.y };
 		const Ray	 ray = reinterpret_cast<const Camera&>(params->Camera).GetRay(uv);
